feat(app): Added required attributes indicator
Closes #62
