+++ apps/app/app/admin/directories/[entityType]/@list/EntitiesTable.tsx
@@ -1,4 +1,4 @@
-import { getEntitiesRaw } from '@gredice/storage';
+import { getAttributeDefinitions, getEntitiesRaw } from '@gredice/storage';
 import { Table } from '@signalco/ui-primitives/Table';
 import { Typography } from '@signalco/ui-primitives/Typography';
 import Link from 'next/link';
@@ -6,8 +6,11 @@ import { Add } from '@signalco/ui-icons';
 import { createEntity } from '../../../../(actions)/entityActions';
 import { ServerActionButton } from '../../../../../components/shared/ServerActionButton';
 import { KnownPages } from '../../../../../src/KnownPages';
+import { Tooltip, TooltipContent, TooltipTrigger } from '@signalco/ui-primitives/Tooltip';
+import { cx } from '@signalco/ui-primitives/cx'
 
 export async function EntitiesTable({ entityTypeName }: { entityTypeName: string }) {
+    const definitions = await getAttributeDefinitions(entityTypeName);
     const entities = await getEntitiesRaw(entityTypeName);
     const createEntityBound = createEntity.bind(null, entityTypeName);
 
@@ -16,18 +19,45 @@ export async function EntitiesTable({ entityTypeName }: { entityTypeName: string
             <Table.Header>
                 <Table.Row>
                     <Table.Head>Naziv</Table.Head>
+                    <Table.Head className='w-20' />
                 </Table.Row>
             </Table.Header>
             <Table.Body>
-                {entities.map(entity => (
-                    <Table.Row key={entity.id}>
-                        <Table.Cell>
-                            <Link href={KnownPages.DirectoryEntity(entityTypeName, entity.id)}>
-                                <Typography>{entity.attributes.find(a => a.attributeDefinition.name === 'name')?.value ?? `${entity.entityType.label} ${entity.id}`}</Typography>
-                            </Link>
-                        </Table.Cell>
-                    </Table.Row>
-                ))}
+                {entities.map(entity => {
+                    const numberOfRequiredAttributes = definitions.filter(d => d.required).length;
+                    const notPopulatedRequiredAttributes = definitions.filter(d => d.required && !entity.attributes.some(a => a.attributeDefinitionId === d.id));
+                    const progress = ((numberOfRequiredAttributes - notPopulatedRequiredAttributes.length) / numberOfRequiredAttributes) * 100;
+                    return (
+                        <Table.Row key={entity.id}>
+                            <Table.Cell>
+                                <Link href={KnownPages.DirectoryEntity(entityTypeName, entity.id)}>
+                                    <Typography>{entity.attributes.find(a => a.attributeDefinition.name === 'name')?.value ?? `${entity.entityType.label} ${entity.id}`}</Typography>
+                                </Link>
+                            </Table.Cell>
+                            <Table.Cell>
+                                <Link href={KnownPages.DirectoryEntity(entityTypeName, entity.id)}>
+                                    {/* Progress bar displaying status of populated required attribute values */}
+                                    <Tooltip delayDuration={250}>
+                                        <TooltipTrigger>
+                                            <div className='py-2 px-1'>
+                                                <div className='h-1 bg-gray-200 rounded-full overflow-hidden w-12'>
+                                                    <div
+                                                        className={cx('h-full', progress <= 99.99 ? 'bg-red-400' : 'bg-green-500')}
+                                                        style={{ width: `${progress}%` }} />
+                                                </div>
+                                            </div>
+                                        </TooltipTrigger>
+                                        <TooltipContent>
+                                            {notPopulatedRequiredAttributes.length === 0
+                                                ? 'Svi obavezni atributi su ispunjeni'
+                                                : `Manjak obaveznih atributa: ${notPopulatedRequiredAttributes.map(a => a.label).join(', ')}`}
+                                        </TooltipContent>
+                                    </Tooltip>
+                                </Link>
+                            </Table.Cell>
+                        </Table.Row>
+                    );
+                })}
                 <Table.Row>
                     <Table.Cell className='p-0'>
                         <ServerActionButton
